+++ src/oph_core.h
@@ -1,6 +1,6 @@
 /*
     Ophidia Primitives
-    Copyright (C) 2012-2017 CMCC Foundation
+    Copyright (C) 2012-2018 CMCC Foundation
 
     This program is free software: you can redistribute it and/or modify
     it under the terms of the GNU General Public License as published by
@@ -30,7 +30,7 @@
 typedef enum oph_hier { INVALID_HIER, OPH_ALL, OPH_WEEK, OPH_MONTH, OPH_YEAR } oph_hier;
 typedef enum oph_type { INVALID_TYPE, OPH_INT, OPH_LONG, OPH_SHORT, OPH_BYTE, OPH_FLOAT, OPH_DOUBLE, OPH_COMPLEX_INT, OPH_COMPLEX_LONG, OPH_COMPLEX_FLOAT, OPH_COMPLEX_DOUBLE } oph_type;
 typedef enum oph_oper { INVALID_OPER, OPH_COUNT, OPH_MAX, OPH_MIN, OPH_SUM, OPH_AVG, OPH_STD, OPH_VAR, OPH_CMOMENT, OPH_ACMOMENT, OPH_RMOMENT, OPH_ARMOMENT, OPH_QUANTILE, OPH_ARG_MAX,
-	OPH_ARG_MIN
+	OPH_ARG_MIN, OPH_AVG_REL
 } oph_oper;
 typedef enum oph_ma_oper { INVALID_MA_OPER, OPH_SMA, OPH_EWMA } oph_ma_oper;
 
@@ -197,9 +197,13 @@ int core_oph_convert(oph_stringPtr byte_array, void *result);
 
 // Return (in byte array form) the sum of the scalar number with every element of byte_array. Cast scalar to byte_array elements type
 int core_oph_sum_scalar(oph_stringPtr byte_array, double scalar, char *result);
+int core_oph_sum_scalar_multi(oph_multistring * byte_array, double scalar, oph_multistring * result, int id);
 
 // Return (in byte array form) the multiplication of the scalar number with every element of byte_array. Cast scalar to byte_array elements type
 int core_oph_mul_scalar(oph_stringPtr byte_array, double scalar, char *result);
+int core_oph_mul_scalar_multi(oph_multistring * byte_array, double scalar, oph_multistring * result, int id);
+
+int core_oph_rel_scalar_multi(oph_multistring * byte_array, double scalar, oph_multistring * result, int id);
 
 /*------------------------------------------------------------------|
  |               Single element functions (END)                     |