+++ src/oph_value_to_bin.c
@@ -1,6 +1,6 @@
 /*
     Ophidia Primitives
-    Copyright (C) 2012-2017 CMCC Foundation
+    Copyright (C) 2012-2018 CMCC Foundation
 
     This program is free software: you can redistribute it and/or modify
     it under the terms of the GNU General Public License as published by
@@ -41,6 +41,7 @@ my_bool oph_value_to_bin_init(UDF_INIT * initid, UDF_ARGS * args, char *message)
 		strcpy(message, "ERROR: Wrong arguments to oph_value_to_bin function");
 		return 1;
 	}
+	args->arg_type[2] = DECIMAL_RESULT;
 
 	initid->ptr = NULL;
 
@@ -51,15 +52,13 @@ void oph_value_to_bin_deinit(UDF_INIT * initid)
 {
 	//Free allocated space
 	if (initid->ptr) {
-		free((oph_string *) initid->ptr);
+		free(initid->ptr);
 		initid->ptr = NULL;
 	}
 }
 
 char *oph_value_to_bin(UDF_INIT * initid, UDF_ARGS * args, char *result, unsigned long *length, char *is_null, char *error)
 {
-	oph_type out_type;
-
 	if (*error) {
 		*length = 0;
 		*is_null = 0;
@@ -72,80 +71,91 @@ char *oph_value_to_bin(UDF_INIT * initid, UDF_ARGS * args, char *result, unsigne
 		*error = 0;
 		return NULL;
 	}
-	out_type = core_get_type(args->args[1], &(args->lengths[1]));
+
+	oph_type out_type = core_get_type(args->args[1], &(args->lengths[1]));
+	size_t size;
 
 	if (!initid->ptr) {
 		// Check provided output type
 		if (!out_type || (out_type != OPH_BYTE && out_type != OPH_SHORT && out_type != OPH_INT && out_type != OPH_LONG && out_type != OPH_FLOAT && out_type != OPH_DOUBLE)) {
 			pmesg(1, __FILE__, __LINE__, "Invalid type\n");
 			*length = 0;
-			*is_null = 1;
+			*is_null = 0;
 			*error = 1;
 			return NULL;
 		}
-
-		initid->ptr = malloc(sizeof(size_t));
+		size = core_sizeof(out_type);
+		initid->ptr = malloc(size);
 		if (!initid->ptr) {
 			pmesg(1, __FILE__, __LINE__, "Error allocating output string\n");
 			*length = 0;
-			*is_null = 1;
+			*is_null = 0;
 			*error = 1;
 			return NULL;
 		}
-		*initid->ptr = core_sizeof(out_type);
-	}
+	} else
+		size = core_sizeof(out_type);
+
+	result = initid->ptr;
+
 	//Decimal_result is passed as a string
 	char buf[256];
-	if (core_strncpy(buf, args->args[2], &(args->lengths[2]))) {
+	if ((args->lengths[2] > 256) || core_strncpy(buf, args->args[2], &(args->lengths[2]))) {
 		pmesg(1, __FILE__, __LINE__, "Invalid string\n");
 		*length = 0;
-		*is_null = 1;
+		*is_null = 0;
 		*error = 1;
 		return NULL;
 	}
 
-	if (args->arg_type[2] == DECIMAL_RESULT) {
-		switch (out_type) {
-			case OPH_INT:{
-					int tmp = (int) strtol(buf, NULL, 10);
-					memcpy(result, (void *) (&tmp), sizeof(int));
-					break;
-				}
-			case OPH_SHORT:{
-					short tmp = (short) strtol(buf, NULL, 10);
-					memcpy(result, (void *) (&tmp), sizeof(short));
-					break;
-				}
-			case OPH_BYTE:{
-					char tmp = (char) strtol(buf, NULL, 10);
-					memcpy(result, (void *) (&tmp), sizeof(char));
-					break;
-				}
-			case OPH_LONG:{
-					long long tmp = strtol(buf, NULL, 10);
-					memcpy(result, (void *) (&tmp), sizeof(long long));
-					break;
-				}
-			case OPH_FLOAT:{
-					float tmp = strtof(buf, NULL);
-					memcpy(result, (void *) (&tmp), sizeof(float));
-					break;
-				}
-			case OPH_DOUBLE:{
-					double tmp = strtod(buf, NULL);
-					memcpy(result, (void *) (&tmp), sizeof(double));
-					break;
-				}
-			default:
-				pmesg(1, __FILE__, __LINE__, "Invalid type\n");
-				*length = 0;
-				*is_null = 1;
-				*error = 1;
-				return NULL;
-		}
-	} else
-		result = (char *) args->args[2];
-	*length = (unsigned long) (*initid->ptr);
+	if (args->arg_type[2] != DECIMAL_RESULT) {
+		pmesg(1, __FILE__, __LINE__, "Invalid type\n");
+		*length = 0;
+		*is_null = 0;
+		*error = 1;
+		return NULL;
+	}
+
+	switch (out_type) {
+		case OPH_INT:{
+				int tmp = (int) strtol(buf, NULL, 10);
+				memcpy(result, (void *) (&tmp), size);
+				break;
+			}
+		case OPH_SHORT:{
+				short tmp = (short) strtol(buf, NULL, 10);
+				memcpy(result, (void *) (&tmp), size);
+				break;
+			}
+		case OPH_BYTE:{
+				char tmp = (char) strtol(buf, NULL, 10);
+				memcpy(result, (void *) (&tmp), size);
+				break;
+			}
+		case OPH_LONG:{
+				long long tmp = strtol(buf, NULL, 10);
+				memcpy(result, (void *) (&tmp), size);
+				break;
+			}
+		case OPH_FLOAT:{
+				float tmp = strtof(buf, NULL);
+				memcpy(result, (void *) (&tmp), size);
+				break;
+			}
+		case OPH_DOUBLE:{
+				double tmp = strtod(buf, NULL);
+				memcpy(result, (void *) (&tmp), size);
+				break;
+			}
+		default:
+			pmesg(1, __FILE__, __LINE__, "Invalid type\n");
+			*length = 0;
+			*is_null = 0;
+			*error = 1;
+			return NULL;
+	}
+
+	*length = (unsigned long) size;
 	*error = 0;
 	*is_null = 0;