+++ HISTORY.md
@@ -1,3 +1,30 @@
+
+## v1.3.0 - 2018-06-18
+
+### Added:
+
+- New primitive oph_sequence [#7](https://github.com/OphidiaBigData/ophidia-primitives/pull/7)
+
+### Fixed:
+
+- Bug in oph_value_to_bin primitive [#6](https://github.com/OphidiaBigData/ophidia-primitives/issues/6)
+
+### Changed:
+
+- oph_gsl_quantile primitive primitive to return index of the closest item to quantile [#8](https://github.com/OphidiaBigData/ophidia-primitives/pull/8)
+
+## v1.2.0 - 2018-02-16
+
+### Added:
+
+- New primitive oph_replace [#4](https://github.com/OphidiaBigData/ophidia-primitives/pull/4)
+- New primitive oph_normalize [#3](https://github.com/OphidiaBigData/ophidia-primitives/pull/3)
+- New primitive oph_padding
+
+### Fixed:
+
+- Bug in multi-fields primitives [#5](https://github.com/OphidiaBigData/ophidia-primitives/pull/5)
+
 ## v1.1.0 - 2017-07-28
 
 ### Fixed: